+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 
 ### New feature support
 
+#### Cisco Resources
+* span_session (@tomcooperca)
+
 ### Added
 
 * Extend interface with attributes:
@@ -444,4 +447,3 @@ Changelog
 [v1.1.0]: https://github.com/cisco/cisco-network-node-utils/compare/v1.0.1...v1.1.0
 [v1.0.1]: https://github.com/cisco/cisco-network-node-utils/compare/v1.0.0...v1.0.1
 [v1.0.0]: https://github.com/cisco/cisco-network-node-utils/compare/v0.9.0...v1.0.0
-